1.12.21.2

+++ main.py
@@ -49,6 +49,8 @@ async def shellTest(ctx):
 
 @bot.command()
 async def pipInstall(ctx, package):
+	if "aiohttp" in package.lower() or "pynacl" in package.lower():
+		return
 	p = Popen("pip install " + package, stdout=PIPE, close_fds=True, shell=True)
 	lines = []
 	msg = await ctx.send('\u200b')
@@ -75,6 +77,12 @@ async def pipInstall(ctx, package):
 
 @bot.command()
 async def pipUninstall(ctx, package):
+	if author.id not == 231628302152892427:
+		print("nope")
+	if author.id not == 231628302152892427:
+		print("yep")
+	if "aiohttp" in package.lower() or "pynacl" in package.lower():
+		return
 	p = Popen("pip uninstall -y " + package, stdout=PIPE, close_fds=True, shell=True)
 	lines = []
 	msg = await ctx.send('\u200b')
